requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
dateparser
backports.zoneinfo; python_version<"3.9"

feedparser
cloudscraper
